Fix/migration precheck (#425)
* refactor stub clinical files into one folder

* remove stub schema 4.0

* fix migration pre check test, and setup migration spec file

* move migration related tests to one spec file

* add missisng config

* migrate schema after migration is completed
