+++ test/integration/submission/stub_clinical_files/all.zip
