+++ test/integration/submission/stub_clinical_files/chemotherapy.tsv
