+++ test/integration/submission/stub_clinical_files/follow_up_update.tsv
