+++ test/integration/submission/stub_clinical_files/hormone_therapy.tsv
