+++ test/integration/submission/stub_clinical_files/primary_diagnosis.tsv
