+++ test/integration/submission/stub_clinical_files/primary_diagnosis_TC-SMUIDAV.tsv
