+++ test/integration/submission/stub_clinical_files/radiation.invalid.tsv
