+++ test/integration/submission/stub_clinical_files/sample_registration.invalid.tsv
