+++ test/integration/submission/stub_clinical_files/specimen.tsv
