+++ test/integration/submission/stub_clinical_files/thisIsARegistration.tsv
