+++ test/integration/submission/stub_clinical_files/thisissample.tsv
