+++ test/integration/submission/stub_clinical_files/treatment.tsv
