+++ test/integration/submission/stub_clinical_files/treatment_update.tsv
